refactor: moving API mocks to support folder
