+++ support/fixtures/get-characters-success.json
