+++ support/fixtures/get-powers-success.json
